Add source mapping to Webpack config
The source mapping dev tool is useful for mapping compressed/transpiled code back to its uncompressed original source. This is useful for debugging -- errors now point to lines in the original ES6/VueJS files instead of in the ES5 compilation.

The gitignore has been modified to ignore resulting mapping files.